backend/game-py/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.0.10
openai>=1.10.0
gymnasium[classic_control]>=0.29.1
numpy>=1.26.0
pillow>=10.2.0


